update test suites
